fix for problem with userDB and bookmarksDB which was caused by changes in kelondroRA in SVN 5376
git-svn-id: https://svn.berlios.de/svnroot/repos/yacy/trunk@5385 6c8d7289-2bf4-0310-a012-ef5d649a1542

+++ source/de/anomic/kelondro/kelondroAbstractRA.java
@@ -51,14 +51,14 @@ abstract class kelondroAbstractRA implements kelondroRA {
     // pseudo-native methods:
     abstract public void readFully(byte[] b, int off, int len) throws IOException;
     abstract public long length() throws IOException;
-    abstract public long available() throws IOException;
+    abstract public int available() throws IOException;
     abstract public void write(byte[] b, int off, int len) throws IOException;
     abstract public void seek(long pos) throws IOException;
     abstract public void close() throws IOException;
 
     // derived methods:
     public byte[] readFully() throws IOException {
-        int a = (int) this.available();
+        int a = this.available();
         if (a <= 0) return null;
         final byte[] buffer = new byte[a];
         this.readFully(buffer, 0, a);
